requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
blinker==1.8.2
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dict-toolbox==5.0.0
distlib==0.3.8
distro==1.9.0
faiss-cpu==1.10.0
filelock==3.13.1
Flask==3.0.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.9.0
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==2.1.2
oauthlib==3.2.2
openai==1.70.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
platformdirs==4.2.0
pop==27.1.0
pop-config==12.0.4
pop-loop==1.1.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.1
RapidFuzz==3.13.0
regex==2024.11.6
rend==7.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tiktoken==0.9.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.1.0
virtualenv==20.25.1
watchdog==3.0.0
Werkzeug==3.1.1
